windows: full ConPTY backend + UTF-8 manifest + Krypton Tier-1 shell hook
Replaces the Windows PTYPlatform stub with a real CreatePseudoConsole-
based backend, widens the cross-platform PTYPlatform interface so the
Windows split-pipe shape fits, and lands a Tier-1 Krypton-lang hook to
prove out the C++↔Krypton interop pattern.

## Interface changes (shared, talk to L if collisions)

include/kterm/pty/PTYPlatform.hpp
  - PTYHandles::masterFd, childPid widened int → long long
    (Windows uses an opaque token; can't fit POSIX-fd semantics).
  - New methods: readData / writeData / closePTY. Linux: thin wrappers
    over POSIX read/write/close. Windows: dispatch through a per-process
    unordered_map<long long, WindowsPTY> guarded by mutex.

src/pty/PTY.cpp
  - Replaced direct POSIX read/write/close with PTYPlatform::readData /
    writeData / closePTY. No behavior change on Linux (same syscalls,
    one level of indirection).
  - m_masterFd type widened to long long.

## Windows-only

platform/windows/PTYPlatform.cpp (was 8-line stub):
  - Targets Win10 1809+ (_WIN32_WINNT 0x0A00, NTDDI_WIN10_RS5).
  - CreatePseudoConsole + STARTUPINFOEXW with
    PROC_THREAD_ATTRIBUTE_PSEUDOCONSOLE_HANDLE + CreateProcessW.
  - ResizePseudoConsole + ClosePseudoConsole on teardown.
  - ReadFile/WriteFile to the child stdin/stdout pipes.
  - resolveShell() tries the Krypton hook (see below), then falls back
    to cmd.exe.

platform/windows/brain-terminal.manifest:
  - activeCodePage = UTF-8 (process ACP flips to 65001 on Win10 1903+
    so byte-pumped output renders Unicode correctly without per-call
    IAT thunks — matches the existing terk.exe.manifest convention).
  - longPathAware = true, dpiAwareness = PerMonitorV2,
    asInvoker (no UAC), supportedOS Win10/11.

CMakeLists.txt:
  - MSVC: writes a generated .rc that embeds the manifest as Win32
    resource 1 (RT_MANIFEST = 24).
  - MinGW/clang-on-Windows: feeds the manifest directly to windres
    via LANGUAGE RC.
  - Linux build path is untouched.

## Krypton Tier-1 PoC (Windows-only)

If kr.exe is on PATH and %APPDATA%\brain-terminal\setup.ks exists,
resolveShell() runs it (CreateProcess + capture stdout + 5-second
bounded wait) and uses the script's last stdout line as the shell
command. No runtime link against krypton_rt.dll; pure CreateProcess
interop. Falls back silently if Krypton isn't installed.

Files:
  - krypton/setup_shell.ks: example shell-picker that detects whether
    we're in a Krypton repo (uses k:env + k:fsx) and picks kr.exe /
    pwsh / cmd / bash / zsh accordingly.
  - KRYPTON_INTEGRATION.md: menu of integration tiers (config, themes,
    status providers, embedded REPL, partial GUI in Krypton).

## Handoff to L

handoffs/handoff_w2l_winport_pty_iface.md tracks the shared touchpoints
and asks for a glance at PTY.hpp/PTYPlatform.hpp + src/pty/PTY.cpp +
the new CMake Windows branches before either of us pushes more.

+++ include/kterm/pty/PTY.hpp
@@ -19,7 +19,10 @@ public:
     void resize(int cols, int rows);
 
 private:
-    int m_masterFd;
+    // Opaque handle from PTYPlatform::createPTY(). On Linux this is the
+    // POSIX master PTY fd; on Windows it's a token into the ConPTY map.
+    // Always go through PTYPlatform for I/O — never POSIX read/write/close.
+    long long m_masterFd;
     OutputCallback m_callback;
 
     void readLoop();
